Replace newline char to HTML line breaks in posts
When a user sends multiple lines post, server will save newline char
to database when it views on web it should be translated to HTML
line breaks.

It results in user see multiple lines posts to be one line on browser.

Add the function that replaces each newline character to HTML line
breaks let users can write multiple lines in a post.